Refactored cucumber steps into separate files
no refs

Refactored cucumber steps into separate files, grouped by either feature or
functionality (for generic steps). `stepdefs.js` has been renamed to `index.js`
and contains the hooks to execute as well as any global setup
